First database implementation
This commit is the result of  all the database implementation

# Constraints on courses
A course is associated with a (unique) room and a (unique) subject
This makes it possible to identify two types of constraints:

- Spatial constraints, a room cannot accommodate two courses simultaneously
- The constraints of availability of actors

## Spatial constraints on rooms

As the room data is stored in the slot, an exclude constraint can be used to prevent insertion according to certain constraints.
Slot overlap can be checked by the && operator. This constraint is controlled by a GiST index (btree_gist postgres extension)
```postgres
EXCLUDE USING gist(room with =,
                 timrange with &&)
)
```

## Constraints on actor availability

A student can belong to several classes (e.g. a double curriculum). However, a group cannot be in several places at the same time
We therefore need to control this when associating a group with a slot. The association class forces us to use a trigger to cross-reference the information.

```postgres
CREATE OR REPLACE FUNCTION public.check_date_overlap_trigger_function_group_slot()
         RETURNS trigger
         LANGUAGE plpgsql
       AS $function$
       begin
       IF EXISTS(select from slots  s
           where id=new.id_slot--- we select the timerange from the id in the query
            and exists (
                    select from group_slot gs --- then we check whether there exists an overlapping one
                    join slots s2 on gs.id_slot=s2.id
                    where gs.id_group = new.id_group
                    and s2.id <> new.id_slot --- we don't want to compare the slot specified in the query to itself
                    and s.timerange && s2.timerange  for share --- we use for share to prevent any modification while the trigger is run.
                       )
       for share)

       THEN
            RAISE unique_violation USING MESSAGE = 'Overlapping timeranges for group : ' || new.id_group;
       END IF;
   RETURN NEW;
       end
       $function$;

CREATE TRIGGER ON group_slot BEFORE INSERT OR UPDATE
FOR EACH ROW
EXECUTE FUNCTION check_date_overlap_trigger_function_group_slot();
```

The previous constraint still applies to lecturers of a course, and we just need to rename the tables in the above

## slot modification constraint

Finally, we need to ensure that actors are still available when a slot is modified.

```postgres
--- TO BE TESTED
```