Unify all id gneration in a single sequence
This commit unifies ID generation in one single sequence in order to
facilitate the future authentication. It will be easier if each entity
has an unique ID source. The DB script is updated accordingly.

Also, the check on slot update for a non-conflicting timerange is done
in the trigger instead of the function to avoid a useless run.